+++ docs/release_notes.rst
@@ -4,6 +4,13 @@ Release notes
 =============
 
 
+* Modified the model to fit the angular coordinate for three-pixel events eta reconstruction.
+* Pull requests merged and issues closed:
+
+  - https://github.com/lucabaldini/hexsample/pull/91
+
+
+
 Version 0.13.1 (2026-01-30)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~